+++ paper.bib
@@ -47,7 +47,7 @@
 }
 
 @online{RubiOrg,
-  author = {The Rubi Organization},
+  author = {{The Rubi Organization}},
   title = {Rule-based Integration -- An extensive system of symbolic integration rules},
   year = 2018,
   url = {https://rulebasedintegration.org/},
@@ -55,7 +55,7 @@
 }
 
 @online{RubiOrgGithub,
-  author = {The Rubi Organization},
+  author = {{The Rubi Organization}},
   title = {Mathematica Test Suite},
   year = 2018,
   url = {https://github.com/RuleBasedIntegration},
@@ -63,7 +63,7 @@
 }
 
 @online{RubiMma,
-  author = {The Rubi Organization},
+  author = {{The Rubi Organization}},
   title = {Mathematica implementation of Rubi},
   year = 2018,
   url = {https://github.com/RuleBasedIntegration/Rubi},
@@ -71,7 +71,7 @@
 }
 
 @online{RubiMmaTestSuite,
-  author = {The Rubi Organization},
+  author = {{The Rubi Organization}},
   title = {Mathematica Test Suite},
   year = 2018,
   url = {https://github.com/RuleBasedIntegration/MathematicaSyntaxTestSuite},
