### 2018-10-31 6a3c0c7
Update paper.md
