### 2018-11-06 #1054 @danielskatz
Sorry for the delay - I missed this earlier.

@halirutan - what is the specific software you are submitting (which repository/repositories)?

Also, if you have any suggestions for potential reviewers then please mention them here in this thread. In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission.
