### 2018-11-06 #1054 @halirutan
@danielskatz Thanks for your response. The specific OSS we want to submit for publishing is the Mathematica implementation for the Rubi integrator which lives here

https://github.com/RuleBasedIntegration/Rubi

However, since this is a larger endeavor we have created a GitHub organization where all repositories live under [github.com/RuleBasedIntegration](https://github.com/RuleBasedIntegration). That was the logical decision as we provide the large database of over 70000 test-integrals for several languages and e.g. the 6600+ integration rules as PDF as well and we wanted a better separation from the implementation. We have provided links to the website and all important repositories in the manuscript.

As reviewers, someone with knowledge of Mathematica and a an understanding of symbolic integration would be good. This narrows down the list pretty much and I suggest

rljacobson
wspr
acolum

If rljacobson is who I think he is, then I do know him from some discussions if that matters. If this is a criterion for exclusion, then I hope he can be at least the 2nd reviewer. He is an expert in both Mathematica and mathematics and I am certain that his suggestions will help to improve the publication.
