### 2018-11-07 #1054 @danielskatz
@rljacobson, @wspr, @acolum - Would you be willing to review this submission?
