### 2018-11-07 #1054 @acolum
Yes, I'd be willing to review this submission. This would be my first time reviewing for JOSS, so I'd appreciate some guidance.
