### 2018-11-09 #1054 @rljacobson
I am happy to review this submission. Like @acolum, this would be my first time reviewing for JOSS, but I fully support its mission and look forward to participating!

For the record, I am [Robert Jacobson, Assistant Professor of Mathematics at Roger Williams University](https://www.rwu.edu/academics/schools-and-colleges/fssns/faculty/robert-jacobson). I hold a PhD in mathematics and have a significant background in computer science. Having used Mathematica professionally for many years and even built compiler tools for the Mathematica programming language, I think it is fair to call me an expert in Mathematica.
