### 2018-11-09 #1073 @whedon
[REVIEW]: Rule-based Integration -- An extensive system of symbolic integration rules
**Submitting author:** @halirutan (<a href="http://orcid.org/0000-0003-0361-9063">Patrick Scheibe</a>)
**Repository:** <a href="https://github.com/RuleBasedIntegration/JOSS-Publication" target ="_blank">https://github.com/RuleBasedIntegration/JOSS-Publication</a>
**Version:** 4.16.0.4

- [x] **Version:** Does the release version given match the GitHub release (4.16.0.4)?

- [x] **Version:** Does the release version given match the GitHub release (4.16.0.4)?
