### 2018-11-09 #1073 @danielskatz
:wave: @acolum and @rljacobson - please go ahead and get started.  Read the opening comment in this issue carefully, and also take a look at the [JOSS reviewer guidance](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).  In brief, you perform your review and check the boxes above when you are satisfied with an item, and open issues in the [paper repository].  (https://github.com/RuleBasedIntegration/JOSS-Publication) or [code repository](https://github.com/RuleBasedIntegration/Rubi) as needed.

Note that JOSS normally assumes the paper is in the same repository as the code, so the repository link in the compiled paper (see the comment above, or a later one if changes are made and the paper is recompiled) will have to be changed by the JOSS editors manually before final acceptance.

If you have any questions, please let me know.
