### 2018-11-10 #1073 @rljacobson
@halirutin @danielskatz Just to clarify, my understanding is that the authors are intending to include in their submission the entire Rubi ecosystem, at least as it relates to its Mathematica implementation, which would include all of the following:

1. The https://github.com/RuleBasedIntegration/Rubi repository
2. The Mathematica test suite, https://github.com/RuleBasedIntegration/MathematicaSyntaxTestSuite
3. The documentation, which presently lives primarily at https://rulebasedintegration.org.
4. The catalog of integration rules, which is housed here: https://github.com/RuleBasedIntegration/IntegrationRules

Is this correct?

If this submission is only a subset of the items I list above, I strongly recommend that the authors revise the submission to include (at least) everything in that list, as those items are all quite entertwined and would make a much more substantial contribution and impact to the field if the submission successfully passes review.
