### 2018-11-10 #1073 @halirutan
@rljacobson You are correct. I had a bit of a struggle while preparing the manuscript, as JOSS seems to be tailored to the "one repository" software written in an open language like python. Rubi is a bit different because the general idea of putting together a large set of integration rules that allow for symbolic integration is at least as important as its implementation. So yes, the Rubi ecosystem consists of the

- integration rules that are available as Mathematica notebooks and rendered PDF files
- it's implementation in a Mathematica package which takes care of setting up the integration rules and wraps everything in a nice user-interface.
- Rubi's test-suite with 70k+ integrals and solution for several CAS languages
- Highlevel user documentation, installation guide, links to CAS comparisons, and other information on [rulebasedintegration.org](https://rulebasedintegration.org/)

Every little part of this is MIT-licenced and freely available. For the JOSS publication, we wanted to present the Mathematica implementation as it is a real program like the other publications on JOSS.
Since we linked in its `README.md` to the Rubi website, it should be sufficient for the interested reader to get all information.

Nevertheless, in the summary manuscript, we tried to present the whole idea and explain important parts of the interconnections. Therefore, the outline reads as

- General importance of symbolic integrators
- Rubi and its development as Mathematica package with integration rules that are repeatedly applied to find antiderivatives
- Details about features and the rigorous test-suite
- Other software packages, that already use Rubi as integrator (which proves the point of being "general" as none of these projects use Mathematica)

> I strongly recommend that the authors revise the submission to include (at least) everything in that list

In the manuscrip we have already included references to 1., 2., and 3. The "Integration Rules" are part of the Rubi Mathematica package and they are explicitely mentioned in the [Repository structure](https://github.com/RuleBasedIntegration/Rubi#general) part of the `README`. The "catalog of integration rules" you mentioned in 4. is in my option not necesarry here, because we can assume that a user reading this manuscript has Mathematica installed and is better off just opening the notebooks.

Would you agree or did I misunderstood your comment?

@danielskatz Is it possible to switch to a numbers for referencing like this [1]? We cite several of our repositories and the website and referencing by (The Rubi Organization, 2018d) doesn't look so good. Also, is it possible to adjust the size of the inserted image?
