### 2018-11-11 #1073 @rljacobson
@halirutan I read your submitted paper just as you describe it, so from my perspective I do not think a revision of the paper is necessary. But as you say, the assumption in the JOSS boilerplate language of a single repository structure does not quite fit this particular application, so I want to make sure we are all on the same page.

@danielskatz A quality review of this software may take longer than is typical for JOSS. First, this software project is quite large. Second, if the functionality and performance claims of this project are verified, it would represent a significant advance in the current state of the art of symbolic computer algebra. The potential impact of this technology recommends a careful peer review, especially with respect to the items _Functionality, Performance_, and _Automated Tests_ of the Review Checklist. These items are also the largest components of this project. I will do my best to be expedient, but I want to communicate to you the size of the review task up front. To be clear, these comments should not be interpreted as a prejudgment of the submission, nor are they intended to in any way disparage the submitting authors, both of whom are well known and respected in their respective subfields for the quality of their work.
