### 2018-11-11 #1073 @danielskatz
👋 @arfon - can you answer these questions about the paper compilation?

> Is it possible to switch to numbers for referencing like this [1]? We cite several of our repositories and the website and referencing by (The Rubi Organization, 2018d) doesn't look so good. Also, is it possible to adjust the size and position of the inserted image?
