### 2018-11-11 397a8f6
Fix references and image
