+++ paper.bib
@@ -50,13 +50,13 @@
   author = {{The Rubi Organization}},
   title = {Rule-based Integration -- An extensive system of symbolic integration rules},
   year = 2018,
-  url = {https://rulebasedintegration.org/},
+  url = {https://rulebasedintegration.org},
   urldate = {2018-10-25}
 }
 
 @online{RubiOrgGithub,
   author = {{The Rubi Organization}},
-  title = {Mathematica Test Suite},
+  title = {The {R}ubi {GitHub} {O}rganization},
   year = 2018,
   url = {https://github.com/RuleBasedIntegration},
   urldate = {2018-10-25}
@@ -64,7 +64,7 @@
 
 @online{RubiMma,
   author = {{The Rubi Organization}},
-  title = {Mathematica implementation of Rubi},
+  title = {Mathematica {I}mplementation of {R}ubi},
   year = 2018,
   url = {https://github.com/RuleBasedIntegration/Rubi},
   urldate = {2018-10-25}
@@ -72,7 +72,7 @@
 
 @online{RubiMmaTestSuite,
   author = {{The Rubi Organization}},
-  title = {Mathematica Test Suite},
+  title = {Rubi {M}athematica {T}est {S}uite},
   year = 2018,
   url = {https://github.com/RuleBasedIntegration/MathematicaSyntaxTestSuite},
   urldate = {2018-10-25}
@@ -86,17 +86,16 @@
  month = jan,
  keywords = {Python, Computer algebra system, Symbolics},
  volume = 3,
- pages = {e103},
+ pages = {103},
  journal = {PeerJ Computer Science},
  issn = {2376-5992},
- url = {https://doi.org/10.7717/peerj-cs.103},
  doi = {10.7717/peerj-cs.103}
 }
 
 @manual{Sage,
   Key          = {SageMath},
   Author       = {{The Sage Developers}},
-  Title        = {{S}ageMath, the {S}age {M}athematics {S}oftware {S}ystem ({V}ersion x.y.z)},
+  Title        = {{S}ageMath, the {S}age {M}athematics {S}oftware {S}ystem},
   note         = {{\tt http://www.sagemath.org}},
   Year         = {2018},
 }
