### 2018-11-12 #1073 @halirutan
@arfon Locally, it is easily possible to make references appear as [1] simply by setting a different citation style. This can be done inside the [header of paper.md](https://raw.githubusercontent.com/RuleBasedIntegration/JOSS-Publication/master/paper.md) with e.g.

```
csl: transactions-on-mathematical-software.csl
```

I included the style file in the repo and, while this works locally with a similar call to pandoc that  @whedon uses

```
pandoc -o test.pdf --filter pandoc-citeproc paper.md
```

it doesn't work here. I'm not sure why. I looked over the whedon code and it seems you clone the repo to a tmp directory and build it. Maybe something with the latex-template you are using interferes.
