### 2018-11-12 #1073 @arfon
> @arfon Locally, it is easily possible to make references appear as [1] simply by setting a different citation style. This can be done inside the header of paper.md with e.g.

Hi @halirutan - I'm afraid we don't offer customizations like this sorry. I realize it's theoretically possible, however, we want to stick to a single citation style.
