### 2018-11-19 #1073 @danielskatz
👋 @acolum & @rljacobson - it looks like both reviews are making progress.  I'm just checking in to see if either of you need anything.  If you're just working through this, great, but if you are stuck somehow, please let me know.
