### 2018-11-21 #1073 @halirutan
@arfon

> I'm afraid we don't offer customizations like this sorry.

No problem at all. I just wanted to make sure I didn't miss something. I will wait what the reviewers think. Maybe I'm converting the "repository citations" to footnotes with links if this is more appropriate. Thanks for looking into this.
