### 2018-11-26 #1073 @rljacobson
> I'm just checking in to make sure everything is going ok - let me know if it's not.

Yep, everything is going well. Thanks!
