### 2018-12-09 #1073 @rljacobson
Doing ok, @danielskatz. Thanks for checking in.
