### 2018-12-10 RuleBasedIntegration/JOSS-Publication#8 @rljacobson
Messages for incorrect usages of Stats, Step(s), etc.
This is a feature request, not a reviewer critique.

I really like the `Int::noShowSteps` message. It even tells me how to fix the problem. Perhaps it would be more clear to say explicitly what "this function" is—obviously it's not the `Int` function. Still, it's a really nice touch.

I found it really easy to accidentally use `Steps`, `Stats`, etc., incorrectly because of my fat fingers. It would be nice to have more messages for incorrect usages like `Stats[Int[x], x]` or `Steps[Int[x], x]`. This one is curious:

```mma
In[8]:= Steps[Int[x, RubiPrintInformation->False]]

Out[8]= Steps[Int[x,RubiPrintInformation->False],4096]
```
