### 2018-12-12 #1073 @rljacobson
@danielskatz I have completed my review.

The only open issue relating to the reviewers' checklist is a [recommendation](https://github.com/RuleBasedIntegration/JOSS-Publication/issues/9) to clarify the first steps for new contributors. However, if you read the details of the issue, you will see that the submission has in many ways substantially met that guideline. It is my opinion that this issue should not prevent the paper's publication as it is, which publication I heartily recommend.

The other "issues" that remain open are only suggestions or what amount to feature requests—they do not reflect my evaluation in my role as reviewer of the paper's fitness for publication.

@halirutan @danielskatz I apologize for the long review period. It took some time to substantially confirm the functionality and performance claims of the software. In the words of Carl Sagan, "Extraordinary claims require extraordinary evidence." The software claims to significantly outperform the state of the art in symbolic integration. I have verified this claim on the project's test suite as well as my own suite of tests. It is quite remarkable.
