### 2018-12-12 #1073 @danielskatz
@halirutan - you might want to respond to some of @rljacobson's recommendations before we publish, perhaps at least the [contributor guidelines](https://github.com/RuleBasedIntegration/JOSS-Publication/issues/9) one.  Please do so, and when you are ready to publish, make an up to date repository in figshare or zenodo or similar, and post the DOI here in a comment.
