### 2018-12-12 RuleBasedIntegration/JOSS-Publication#9 @halirutan
I have put a small "How to contribute" section directly on the [first page of the wiki](https://github.com/RuleBasedIntegration/Rubi/wiki). As you pointed out, the best way is to first talk to us on Gitter and I have pointed this out.

In addition, I started something I initially saw when working on JabRef: Tagging specific issue in the bug-tracker as "help-wanted" or "good first issue" if the problem is easy enough so that a new contributor can directly work on it.

So right now, I [have created new issues](https://github.com/RuleBasedIntegration/Rubi/labels/help%20wanted) that cover the things that should be a goal for 2019 and where OSS contributors could really help us.
