### 2018-12-12 RuleBasedIntegration/JOSS-Publication#8 @halirutan
Good catch! I've turned this into a [Rubi issue](https://github.com/RuleBasedIntegration/Rubi/issues/19) and I believe both things can be fixed without much work.
