### 2018-12-13 RuleBasedIntegration/JOSS-Publication#8 @halirutan
I already fixed this.
