### 2018-12-13 #1073 @halirutan
Let me thank both reviewers for taking their time to look over Rubi.

To give a short summary of @rljacobson suggestions:

1. I made clear in the README and the wiki how new contributors should start. Basically, it's "please talk to us on Gitter first", but to give some further guidance ...
2. I have created issues for improvements that range from trivial to hard and tagged them as [help wanted](https://github.com/RuleBasedIntegration/Rubi/issues?q=is%3Aissue+is%3Aopen+label%3A%22help+wanted%22). This is also pointed out in the Wiki and a good starting point for new contributors.
3. I improved the error message when certain functions are used incorrectly as suggested [here](https://github.com/RuleBasedIntegration/JOSS-Publication/issues/8).
4. The remaining two suggestions have been turned into issues in the bug tracker so that contributors can work on them. Both issues would be a perfect fit since they are rather trivial but require some time. I still haven't lost the hope that people start contributing.

With these changes, I created a new release version **4.16.0.5** and an used Zenodo to get a DOI:

[![DOI](https://zenodo.org/badge/135434803.svg)](https://zenodo.org/badge/latestdoi/135434803)

That should be all. Finally, thank you @danielskatz for guiding us along the way.
