### 2018-12-13 RuleBasedIntegration/JOSS-Publication#9 @rljacobson
Sounds reasonable.
