### 2018-12-16 #1073 @danielskatz
👋 @arfon - it looks like something didn't work here - the pdf in the PR looks good, but it doesn't seem to be in http://joss.theoj.org/papers/10.21105/joss.01073 correctly.  https://www.theoj.org/joss-papers/joss.01073/10.21105.joss.01073.pdf gives a 404 error.
