### 2018-12-17 #1073 @arfon
@danielskatz - those links seem to work for me now. Perhaps GitHub was being a little slow?
