### 2018-12-17 #1073 @danielskatz
@halirutan - your paper is now accepted into JOSS ⚡️🚀💥
